# docs/swagger.yaml
basePath: /
definitions:
  avito.Operation:
    properties:
      dateTime:
        type: string
      operation_type:
        type: string
      segment_name:
        type: string
      userId:
        type: integer
    type: object
  avito.Segment:
    properties:
      segment_name:
        type: string
    required:
    - segment_name
    type: object
  avito.SegmentInSwagger:
    properties:
      segmentName:
        type: string
    required:
    - segmentName
    type: object
  avito.UserIdOnly:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
  avito.UserInfo:
    properties:
      end:
        type: string
      start:
        type: string
      userId:
        type: integer
    required:
    - end
    - start
    - userId
    type: object
  avito.UserList:
    properties:
      added:
        type: string
      expiry:
        type: string
      segment_names_add:
        items:
          type: string
        type: array
      segment_names_remove:
        items:
          type: string
        type: array
      userId:
        type: integer
    required:
    - segment_names_add
    - segment_names_remove
    - userId
    type: object
  avito.UserSegmentWithId:
    properties:
      segmentName:
        type: string
      userId:
        type: integer
    required:
    - segmentName
    - userId
    type: object
  handler.customError:
    properties:
      message:
        type: string
    type: object
  handler.getAllUserOperationsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/avito.Operation'
        type: array
    type: object
  handler.getAllUserSegmentsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/avito.Segment'
        type: array
    type: object
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Application task for backend trainee position
  title: Avito Trainee Application
  version: "1.0"
paths:
  /api/getUserSegments:
    post:
      consumes:
      - application/json
      description: Получение всех сегментов пользователя
      operationId: get-user-segments
      parameters:
      - description: id пользователя
        in: body
        name: userId
        required: true
        schema:
          $ref: '#/definitions/avito.UserIdOnly'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllUserSegmentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.customError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.customError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.customError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.customError'
      summary: getUserSegments
      tags:
      - segment
  /api/segments:
    delete:
      consumes:
      - application/json
      description: Удаление сегмента
      operationId: delete-segment
      parameters:
      - description: Имя удаляемого сегмента
        in: body
        name: segmentName
        required: true
        schema:
          $ref: '#/definitions/avito.SegmentInSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.customError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.customError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.customError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.customError'
      summary: deleteSegment
      tags:
      - segment
    post:
      consumes:
      - application/json
      description: Создание нового сегмента
      operationId: create-segment
      parameters:
      - description: Имя нового сегмента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito.Segment'
      produces:
      - application/json
      responses:
        "200":
          description: response
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.customError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.customError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.customError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.customError'
      summary: createSegment
      tags:
      - segment
  /api/userInfo/:
    post:
      consumes:
      - application/json
      description: Получение информации о добавлении и удалении пользователя в сегменты
        во временном промежутке
      operationId: get-user-info
      parameters:
      - description: id пользователя и временной промежуток. Поля end и start являются
          строками и имеют формат записи ДД.ММ.ГГГГ
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/avito.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllUserOperationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.customError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.customError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.customError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.customError'
      summary: getUserInfo
      tags:
      - user
  /api/users/:
    delete:
      consumes:
      - application/json
      description: Удаление пользователя из сегмента
      operationId: delete-user
      parameters:
      - description: id пользователя и название сегмента, из которого его нужно удалить
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/avito.UserSegmentWithId'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.customError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.customError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.customError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.customError'
      summary: deleteUser
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Добавление и удаление пользователя в указанные сегменты
      operationId: create-user
      parameters:
      - description: Поля added и expiry являются необязательными и имеют формат ДД.ММ.ГГГГ.
          segment_names_add - список сегментов, в которые нужно добавить пользователя.
          segment_names_remove - список сегментов, из которых нужно удалить пользователя.
          Названия сегментов в списках перечисляются через запятую
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito.UserList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.customError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.customError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.customError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.customError'
      summary: createUser
      tags:
      - user
swagger: "2.0"